requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
bcrypt==4.0.1
cachetools==5.3.2
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.7
Django==4.2.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
lxml==5.2.1
MarkupSafe==2.1.5
msal==1.26.0
numpy==1.25.1
oauth2client==4.1.3
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pdfkit==1.0.0
Pillow==10.1.0
protobuf==4.25.1
psycopg2-binary==2.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
